+++ server/proxy.js
@@ -7,8 +7,10 @@ const log = require('pilogger');
 const proxy = require('http2-proxy');
 const geoip = require('./geoip.js');
 
+let app;
+
 function errorHandler(err) {
-  if (err) log.error('Proxy error', err);
+  if (err) log.error('Proxy error', err.code);
 }
 
 function logger(req, res) {
@@ -17,7 +19,7 @@ function logger(req, res) {
   const agentDetails = `OS:'${agent.os.family}' Device:'${agent.device.family}' Agent:'${agent.family}.${agent.major}.${agent.minor}'`;
   const peer = req.socket._peername || {};
   const ip = peer.address || req.socket.remoteAddress;
-  const geo = peer.address ? geoip.get(ip) : {};
+  const geo = ip ? geoip.get(ip) : {};
   const geoDetails = geo.country ? `Geo:'${geo.continent}/${geo.country}/${geo.city}' ASN:'${geo.asn}' Loc:${geo.lat},${geo.lon}` : '';
   const size = res.headers ? (res.headers['content-length'] || res.headers['content-size'] || 0) : 0;
   const client = `${head[':scheme'] || (req.socket.encrypted ? 'https' : 'http')}://${head[':authority'] || req.headers.host}${req.url}`;
@@ -56,6 +58,44 @@ function findTarget(req) {
   return res;
 }
 
+function startMiddleware() {
+  // Use connect as middleware enabler
+  // eslint-disable-next-line global-require
+  const connect = require('connect');
+  app = connect();
+
+  // Enable middleware
+
+  log.info('Enabling Helmet protection');
+  // eslint-disable-next-line global-require
+  const helmet = require('helmet');
+  app.use(helmet());
+
+  // eslint-disable-next-line global-require
+  const limiter = require('./limiter.js').limiter;
+  log.info('Enabling rate limiter:');
+  app.use(limiter);
+
+  /*
+  log.info('Enabling Brotli compression');
+  const shrink = require('shrink-ray-current');
+  app.use(shrink({ useZopfliForGzip: false, brotli: { quality: 4 } }));
+
+  log.info('Enabling rate limiter');
+  const rateLimit = require('express-rate-limit');
+  const limiter = rateLimit({
+    windowMs: 15 * 60 * 1000, // 15 minutes
+    max: 50, // limit each IP to 100 requests per windowMs
+  });
+  app.use(limiter);
+  */
+
+  // Actual proxy calls
+  log.info('Activating reverse proxy');
+  // eslint-disable-next-line no-unused-vars
+  app.use((req, res, next) => proxy.web(req, res, findTarget(req), errorHandler));
+}
+
 function startServer(ssl) {
   // Start Proxy web server
   global.config.http2.key = fs.readFileSync(path.join(__dirname, ssl.Key));
@@ -64,7 +104,8 @@ function startServer(ssl) {
   server.on('listening', () => log.state('Proxy listening:', server.address()));
   server.on('error', (err) => log.error('Proxy error', err.message));
   server.on('close', () => log.state('Proxy closed'));
-  server.on('request', (req, res) => proxy.web(req, res, findTarget(req), errorHandler));
+  // server.on('request', (req, res) => proxy.web(req, res, findTarget(req), errorHandler));
+  server.on('request', app);
   server.listen(global.config.http2.port);
 }
 
@@ -73,10 +114,12 @@ async function init(ssl) {
   redirectSecure();
 
   // Start proxy web server
+  startMiddleware();
   startServer(ssl);
 
   // Log all redirect rules and start redirecting
-  for (const rule of global.config.redirects) log.info(' Rule:', rule);
+  for (const rule of global.config.redirects) log.info(' Rule:', JSON.stringify(rule));
 }
 
 exports.init = init;
+exports.logger = logger;